Fixed escape sequence conflicts & folding slowdown.
- Removed remaps for home/end/pgup/pgdn. They were causing Vim to wait
for codes after <esc>, and so it would wait after switching from
insert/visual/replace into normal. That meant I had to use
ttimeoutlen/timeoutlen just to get Vim to not hang. But that meant that
Vim wouldn't respond to <leader> or wait for input, and that broke half
of the functionality in a serious majority of plugins. Ugh. The buck
stops here. I went back to stock.

- Added FastFold to fix problem where insert mode (and undo/redo in
normal) slows to a crawl on files with folds in them. Apparently, this
little plugin works like a charm.